fine-turnning/unsloth/requirements.txt
unsloth
wandb
python-dotenv